feat(loader): show loader after a specified amount of ms.
affects: webiny-app-ui

Sometimes loader is shown and hidden so fast, that it looks like a glitch. This will enable
component to wait for a brief amount of time (default 50ms), and to show loader after the time has
passed, which means it won't be shown at all if the component was unmounted quickly.